rename catboost_prior -> catboost
